feat: cost/eval analytics dashboard UI (Phase 3); fix bigint sum decode bug
SUM(integer) returns bigint but SUM(bigint) returns numeric in Postgres —
the analytics query summed token counts as ::bigint, which sqlx then
failed to decode into an i64 field (500 on any execution with node data).
Cast to ::integer instead so SUM's bigint result matches.

+++ apps/api/src/analytics.rs
@@ -28,8 +28,8 @@ pub async fn workflow_analytics(
              e.status,
              e.started_at,
              COALESCE(SUM((n.output->>'cost_usd')::double precision), 0) AS total_cost_usd,
-             COALESCE(SUM((n.output->>'input_tokens')::bigint), 0) AS total_input_tokens,
-             COALESCE(SUM((n.output->>'output_tokens')::bigint), 0) AS total_output_tokens,
+             COALESCE(SUM((n.output->>'input_tokens')::integer), 0) AS total_input_tokens,
+             COALESCE(SUM((n.output->>'output_tokens')::integer), 0) AS total_output_tokens,
              AVG((n.output->>'score')::double precision) AS avg_eval_score
          FROM workflow_executions e
          LEFT JOIN workflow_execution_nodes n ON n.execution_id = e.id
